update_prototype_working
updated FinalAdSensor with a folder of that name (bc my stupid Arduino ide cannot handle it without) and installed all libraries, and got the original code running on our prototype, and added a necessary zip library to the repo. Disconnects sometimes but just reupload the code and it should print fine again.